requirements.txt
config==0.5.1
Flask==2.3.3
flask_cors==4.0.0
openai==1.48.0
python-dotenv==0.21.1
torch==2.4.1
pyd
pydantic_settings==2.5.2
numpy==1.23.5